Updated timer channel allocation functions.
Allocation functions if GpPwm, GpTimer and Sct now have consistent API.

+++ include/platform/nxp/gptimer_base.h
@@ -54,9 +54,9 @@ struct GpTimerBase
   uint8_t resolution;
 };
 /*----------------------------------------------------------------------------*/
-int8_t gpTimerAllocateChannel(uint8_t);
-int8_t gpTimerConfigCapturePin(uint8_t, pinNumber, enum pinPull);
-int8_t gpTimerConfigMatchPin(uint8_t, pinNumber);
+int gpTimerAllocateChannel(uint8_t);
+uint8_t gpTimerConfigCapturePin(uint8_t, pinNumber, enum pinPull);
+uint8_t gpTimerConfigMatchPin(uint8_t, pinNumber);
 
 uint32_t gpTimerGetClock(const struct GpTimerBase *);
 /*----------------------------------------------------------------------------*/
